CiviMail - Move angular settings to its own file
